style(gpu_ntt): drive the crate clippy-clean
Clear all 48 gpu_ntt-attributed clippy findings (both --all-targets and
--features bench) mechanically, no logic change:

- drop redundant same-type casts (u32/usize/i32) in launchers, test seed
  helpers, and the bench grid dim;
- rewrite manual modulo checks as is_multiple_of and the L2 clamp as
  checked_div; OR-patterns as ranges (1..=3, 8..=12);
- drop needless borrows/derefs (mod.rs DeviceMatrix, test out_ptr,
  dispatch as_deref_mut) and the needless late-init/return in lde.rs;
- degroup the 0x1de<log_n> host-oracle seeds so digit groups are uniform;
- reflow the dit.rs module-doc line so a leading '+' is no longer parsed
  as a markdown list marker.

+++ gpu/ntt/src/ntt/hypercube.rs
@@ -112,7 +112,7 @@ pub(crate) fn hypercube_evals_to_monomials_3_pass(
                 input,
                 output_matrix_mut,
                 log_n as i32,
-                start_stage as i32,
+                start_stage,
             );
             StridedTilesStagesFunction(ab_hypercube_evals_to_monomials_nonfinal_8_stages_kernel)
                 .launch(&config, &args)?;
